+++ src/main/java/com/zerobase/musinsamonitor/controller/CartController.java
@@ -1,9 +1,8 @@
 package com.zerobase.musinsamonitor.controller;
 
-import com.zerobase.musinsamonitor.dto.CartResponseDto;
-import com.zerobase.musinsamonitor.dto.ProductResponseDto;
+import com.zerobase.musinsamonitor.model.requestdto.CartDeleteRequest;
+import com.zerobase.musinsamonitor.model.responsedto.CartResponseDto;
 import com.zerobase.musinsamonitor.repository.entity.MemberEntity;
-import com.zerobase.musinsamonitor.security.jwt.TokenProvider;
 import com.zerobase.musinsamonitor.service.CartService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,13 +11,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
@@ -37,10 +34,10 @@ public class CartController {
     }
 
     @PostMapping("/cart/delete")
-    public void deleteProductFromCart(@RequestParam("delete_no") String deleteNo,
+    public void deleteProductFromCart(@RequestBody CartDeleteRequest productIds,
         @AuthenticationPrincipal MemberEntity member) {
-        log.info("삭제 : " + deleteNo);
-        cartService.deleteProductFromCart(deleteNo, member.getEmail());
+        log.info("삭제 번호: " + productIds.getProductIds());
+        cartService.deleteProductFromCart(productIds.getProductIds(), member.getEmail());
     }
 
     @GetMapping("/cart")
