+++ src/main/java/com/zerobase/musinsamonitor/model/requestdto/CartDeleteRequest.java
@@ -1,0 +1,19 @@
+package com.zerobase.musinsamonitor.model.requestdto;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Setter
+@Getter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class CartDeleteRequest {
+    private List<Integer> productIds;
+}
